test: enable DataIntegrity integration tests missing the Test suffix
Seven integration tests under dhis-test-web-api/.../controller/dataintegrity
did not match surefire's include patterns (Test*, *Test, *Tests, *TestCase),
so they were never scanned by the integration-test profile and have not run
since they were committed. They extend AbstractDataIntegrityIntegrationTest
like their ~82 siblings; only the file/class name differed.

Rename all seven to end in Test so surefire picks them up:

  DataIntegrityProgramStagesNoPrograms
  DataIntegrityOptionGroupsWithoutOptions
  DataIntegrityUsersCaptureOrgUnitNotInTeiSearchHierarchy
  DataIntegrityUserRolesNoUsers
  DataIntegrityUsersCaptureOrgunitNotInDataView
  DataIntegrityProgramInconsistentlyLinkedTrackedEntityType
  DataIntegrityUserRolesNoAuthorities

Enabling them surfaced that DataIntegrityProgramInconsistentlyLinkedTracked
EntityType built Program instances without the enrollmentCategoryCombo that
became a not-null property in #22698 (added after this test was written but
never run, so it never failed). Build its Programs via the createProgram /
createProgramWithoutRegistration TestBase helpers, which #22698 updated to set
both the category combo and enrollment category combo to the default - the
same pattern #22698 applied to the other DataIntegrity tests. The check under
test (programs_inconsistent_tracked_entity_type) is unchanged and orthogonal
to category combos; the assertions on program type / tracked entity type are
untouched.

Verified locally: mvn -Pintegration-test -Dtest=<7 classes> -> 12 tests, 0
failures, 0 errors; spotless:check passes.

Caught by @david-mackessy in review of #24010.
